pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tax-prep-agent"
version = "0.1.0"
description = "A CLI agent for tax document collection, analysis, and return review"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["tax", "cli", "agent", "ocr", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "PyMuPDF>=1.24.0",
    "sqlcipher3-binary>=0.5.0",
    "keyring>=25.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "boto3>=1.34.0",  # For AWS Bedrock support
    "pillow>=10.0.0",  # For image processing
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
tax-agent = "tax_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/tax_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
